refactor(main): sync image for build
