+++ docs/4.0/docs/build-cloud-image/build-template-intro.md
